refactor(bin): move events to a separate header
